back-end/requirements.txt
attrs==19.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
ecdsa==0.13.2
Flask==1.0.3
Flask-Cors==3.0.8
fpdf==1.7.2
future==0.17.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.7
pipenv==2018.11.26
pyasn1==0.4.5
pyrsistent==0.15.5
python-dotenv==0.10.3
python-jose==3.0.1
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
six==1.12.0
swagger-ui-bundle==0.0.5
urllib3==1.25.3
virtualenv==16.7.8
virtualenv-clone==0.5.3
Werkzeug==0.16.0
